colored lighting

+++ src/workers/chunkMeshingWorker.ts
@@ -130,7 +130,7 @@ workerSelf.onmessage = (event: MessageEvent<ChunkMeshWorkerMessage>): void => {
       const chunk = workerManager.getOrCreateChunk(snapshot.chunkX, snapshot.chunkZ);
       chunk.loadGeneratedBlocks(new Uint8Array(snapshot.blocks));
       chunk.loadGeneratedMetadata(new Uint8Array(snapshot.metadata));
-      chunk.loadLightData(new Uint8Array(snapshot.light));
+      chunk.loadLightData(new Uint16Array(snapshot.light));
       chunk.markClean();
     }
